Create hook to get logo image details
