+++ src/components/experience.js
@@ -1,37 +1,12 @@
 import React from 'react';
-import { graphql, useStaticQuery } from 'gatsby';
+
+import useLogoImages from '../hooks/useLogoImages';
 
 import SectionItem from './SectionItem';
 
-import Img from 'gatsby-image';
-
 function Experience(props) {
   const { experiences } = props;
-
-  // Get SVG logo details using GraphQL
-  const data = useStaticQuery(graphql`
-    {
-      logoImages: allFile(filter: { relativeDirectory: { eq: "logos" } }) {
-        results: edges {
-          logo: node {
-            name
-            id
-            publicURL
-            relativePath
-          }
-        }
-      }
-    }
-  `);
-
-  // Sets the key and values of the dictionary to the logo name and detail respectively
-  let logoImages = {};
-  if (data) {
-    data.logoImages.results.map(item => {
-      const { name: logoName, ...logoDetails } = item.logo;
-      logoImages[logoName] = logoDetails;
-    });
-  }
+  const logoImages = useLogoImages();
 
   const renderJobExperiences = (experiences, logoImages) => {
     // TODO: Sort the experiences array before rendering.
